+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## HEAD
 - `package.json` cleanups.
+- Added `trailingSlash` option with `true` as default value. When set to `false`, all trailing slashes from URLs will be removed.
 - Added `package.json` `main` property pointing to the task file.
 - Added `snazzy` for better `standard` output.
 - Added NPM 5 lockfile.